update product response with category

+++ src/main/java/com/example/lunastreaming/model/CategoryEntity.java
@@ -3,8 +3,6 @@ package com.example.lunastreaming.model;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.time.LocalDateTime;
-
 @Entity
 @Table(name = "category")
 @Data
@@ -15,7 +13,7 @@ public class CategoryEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false)
     private String name;
